requirements.txt
torch
torchvision
transformers
Pillow
flask
requests
